requirements.txt
numpy
pandas
polars
scikit-learn
torch
torchvision
matplotlib
seaborn
tqdm
docopt-ng
scipy
duckdb
ucimlrepo